ButtonManager->UI_Manager
Made ButtonManager a global accessible singleton
Renamed to UI_Manager
Transferred UI code from GameMan
