+++ Assets/Scenes/SampleScene.unity
@@ -588,6 +588,7 @@ GameObject:
   m_Component:
   - component: {fileID: 696226420}
   - component: {fileID: 696226419}
+  - component: {fileID: 696226421}
   m_Layer: 0
   m_Name: Manager
   m_TagString: Untagged
@@ -622,6 +623,18 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &696226421
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 696226418}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7daee55b2bfd4824fae3890f743762aa, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &711826519
 GameObject:
   m_ObjectHideFlags: 0
